Big update from Standalone
Updates the the converter with all the general stuff that has been added
to the Standalone version in the past year or so, including map changes.
Should not include any standalone-specific stuff, or any artifaqcts from
the standalone's own original converted save game.

Specific updates to province history files:
- Removed all instances of "remove_core" in province history file. This
was to cut down on error log spam in the Standalone, but shouldn't have
an effect on converted games.
- New provinces in Africa
- Changed the script for all centers of trade, due to script changes by
Paradox.
- Removed all lines giving provinces to estates at game-start. This was
to solve an issue where removed or empty provinces were being given to
estates, causing errors.
- Changed culture of some Russian provinces from Russian to
Novogorodian.